refactor: move getAddressAccountIndex out of verify address
