+++ packages/cardano-metamask-snap/src/api/cardano__verifyAddress/index.ts
@@ -1,11 +1,12 @@
 import { type JsonRpcRequest } from '@metamask/snaps-sdk';
 
-import { addressTypes, isAddressParams, type AddressParams } from '../address';
-import { cryptoProvider } from '../cryptoProvider';
 import {
-  getPaymentPathAccountIndex,
-  getStakePathAccountIndex,
-} from '../derivationPath';
+  addressTypes,
+  getAddressAccountIndex,
+  isAddressParams,
+  type AddressParams,
+} from '../address';
+import { cryptoProvider } from '../cryptoProvider';
 import { type NetworkId, isNetworkId, networkIds } from '../networkId';
 import { assertIsArray, isRecord } from '../utils';
 import { renderVerifyAddress } from './ui';
@@ -50,44 +51,6 @@ export function assertIsVerifyAddressRequestParams(
       )} `,
     );
   }
-}
-
-/**
- * Addresses which payment derivation path is equal to 1852'/1815'/<accountIndex>'/0/0
- * and stake derivation path is equal to 1852'/1815'/<accountIndex>'/2/0 are considered account addresses.
- *
- * @param params - Address params.
- * @param params.addressType - The type of address.
- * @param params.paymentDerivationPath - The payment derivation path.
- * @param params.stakeDerivationPath - The stake derivation path.
- * @returns The account index if paths account indexes match, null if they don't or if we don't consider it a "standard" address path for the sake of NuFi wallet.
- */
-function getAddressAccountIndex({
-  addressType,
-  paymentDerivationPath,
-  stakeDerivationPath,
-}: VerifyAddressRequestParams[number]['addressParams']) {
-  if (
-    addressType === addressTypes.BasePaymentKeyStakeKey &&
-    paymentDerivationPath &&
-    stakeDerivationPath
-  ) {
-    const paymentAccountAddressIndex = getPaymentPathAccountIndex(
-      paymentDerivationPath,
-    );
-    const stakeAccountAddressIndex =
-      getStakePathAccountIndex(stakeDerivationPath);
-    return paymentAccountAddressIndex === stakeAccountAddressIndex
-      ? stakeAccountAddressIndex
-      : null;
-  }
-  if (addressType === addressTypes.EnterpriseKey && paymentDerivationPath) {
-    return getPaymentPathAccountIndex(paymentDerivationPath);
-  }
-  if (addressType === addressTypes.RewardKey && stakeDerivationPath) {
-    return getStakePathAccountIndex(stakeDerivationPath);
-  }
-  return null;
 }
 
 export const verifyAddress = async ({
